# [BUGFIX] Always cache TypoScript config in FluxService