--- a/Classes/Service/FluxService.php
+++ b/Classes/Service/FluxService.php
@@ -353,17 +353,22 @@ class FluxService implements SingletonInterface {
 	 * @return array
 	 */
 	public function getAllTypoScript() {
-		if (!$this->configurationManager instanceof BackendConfigurationManager) {
-			$typoScript = (array) $this->configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_FULL_TYPOSCRIPT);
-			$typoScript = GeneralUtility::removeDotsFromTS($typoScript);
-			return $typoScript;
+		$pageId = $this->getCurrentPageId();
+		if (FALSE === isset(self::$typoScript[$pageId])) {
+			self::$typoScript[$pageId] = (array) $this->configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_FULL_TYPOSCRIPT);
+			self::$typoScript[$pageId] = GeneralUtility::removeDotsFromTS(self::$typoScript[$pageId]);
+		}
+		return (array) self::$typoScript[$pageId];
+	}
+
+	/**
+	 * @return integer
+	 */
+	protected function getCurrentPageId() {
+		if ($this->configurationManager instanceof BackendConfigurationManager) {
+			return (integer) $this->configurationManager->getCurrentPageId();
 		} else {
-			$pageId = $this->configurationManager->getCurrentPageId();
-			if (FALSE === isset(self::$typoScript[$pageId])) {
-				self::$typoScript[$pageId] = (array) $this->configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_FULL_TYPOSCRIPT);
-				self::$typoScript[$pageId] = GeneralUtility::removeDotsFromTS(self::$typoScript[$pageId]);
-			}
-			return (array) self::$typoScript[$pageId];
+			return (integer) $GLOBALS['TSFE']->id;
 		}
 	}